+++ src/components/ItemsList/ItemsList.js
@@ -5,6 +5,7 @@ import { item } from '../../types';
 import ErrorBoundary from '../ErrorBoundary';
 import ShippingDetails from '../ShippingDetails';
 import Item from '../Item';
+import Content from '../Content';
 import Wrapper from './Wrapper';
 import List from './List';
 
@@ -17,7 +18,9 @@ function ItemsList({ items }) {
             <Item key={item.id} item={item} />
           ))}
         </List>
-        <ShippingDetails />
+        <Content>
+          <ShippingDetails />
+        </Content>
       </Wrapper>
     </ErrorBoundary>
   );
