+++ src/components/PaymentInformation/PaymentMethod.js
@@ -7,7 +7,7 @@ import Wrapper from './Wrapper';
 import PaymentCard from './PaymentCard';
 import CardName from './CardName';
 import Price from './Price';
-import Section from './Section';
+import Section from '../Section';
 
 function hiddenCardNumber(cardNumber) {
   return `XXXX-XXXX-XXXX-${cardNumber.slice(12)}`;
