--- src/components/ShippingDetails/MessageTitle.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import Text from '../Text';
-
-const MessageTitle = styled(Text)`
-  margin: 0;
-`;
-
-export default MessageTitle;
