--- src/components/ShippingDetails/Section.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-import { spacingMedium } from '../../styles/designTokens';
-
-const Section = styled.section`
-  margin: 0 auto ${spacingMedium};
-  width: 100rem;
-`;
-
-export default Section;
