+++ src/components/ShippingDetails/ShippingDetails.js
@@ -4,14 +4,11 @@ import { colorLightGrey } from '../../styles/designTokens';
 import { formatDate } from '../../utils';
 import Shipping from '../Icons/Shipping';
 import Text from '../Text';
-import ArrivalDates from './ArrivalDates';
-import Message from './Message';
-import MessageTitle from './MessageTitle';
-import Section from './Section';
+import ShippingMessage from './ShippingMessage';
 import Wrapper from './Wrapper';
 
 function ShippingDetails() {
-  const arrivalDates = () => {
+  const getArrivalDates = () => {
     const startDate = new Date();
     const endDate = new Date();
     endDate.setDate(endDate.getDate() + 3);
@@ -20,19 +17,21 @@ function ShippingDetails() {
   };
 
   return (
-    <Section>
-      <Text bold as="h3">
-        Item can only be shipped Standard
-      </Text>
+    <>
+      <Text bold>Item can only be shipped Standard</Text>
       <Wrapper>
         <Shipping fill={colorLightGrey} />
-        <Message>
-          <MessageTitle>Ship to Me</MessageTitle>
-          <Text as="span">Arrives</Text>
-          <ArrivalDates>{arrivalDates()}</ArrivalDates>
-        </Message>
+        <ShippingMessage>
+          <p>Ship to Me</p>
+          <p>
+            Arrives{' '}
+            <Text as="span" bold>
+              {getArrivalDates()}
+            </Text>{' '}
+          </p>
+        </ShippingMessage>
       </Wrapper>
-    </Section>
+    </>
   );
 }
