+++ src/components/ShippingDetails/Wrapper.js
@@ -1,8 +1,11 @@
 import styled from 'styled-components';
 
+import { spacingMedium } from '../../styles/designTokens';
+
 const Wrapper = styled.div`
   display: flex;
   align-items: center;
+  margin-bottom: ${spacingMedium};
 `;
 
 export default Wrapper;
